+++ src/debugger/primitivetypes/types.h
@@ -40,6 +40,8 @@ HRESULT CreateICorValue(ICorDebugThread *pThread, CorElementType elemType, void 
 HRESULT CreateICorValue(ICorDebugThread *pThread, EvalHelpers *pEvalHelpers, PrimitiveValue &primValue, ICorDebugValue **ppValue);
 
 HRESULT CalculateUnary(Parser::SyntaxKind kind, const PrimitiveValue &inputValue, PrimitiveValue &outputValue, std::string &output);
+HRESULT CalculateBinary(Parser::SyntaxKind kind, const PrimitiveValue &leftValue, const PrimitiveValue &rightValue,
+                        PrimitiveValue &outputValue, std::string &output);
 
 } // namespace dncdbg::PrimitiveTypes