+++ src/managed/interop.h
@@ -209,8 +209,6 @@ class Interop
                                               std::vector<AsyncAwaitInfoBlock> &AsyncAwaitInfo, uint32_t *ilOffset);
     static HRESULT GetLocalConstants(void *pSymbolReaderHandle, mdMethodDef methodToken, uint32_t ilOffset,
                                      void **data, int32_t &constantCount);
-    static HRESULT Calculation(void *firstOp, int32_t firstType, void *secondOp, int32_t secondType, int32_t operationType,
-                               int32_t &resultType, void **data, std::string &errorText);
     static void *AllocString(const std::string &str);
     static HRESULT StringToUpper(std::string &String);
     static BSTR SysAllocStringLen(int32_t size);
@@ -251,7 +249,6 @@ class Interop
     using ResolveBreakPointsDelegate = RetCode (*)(void *, int32_t, void *, int32_t, int32_t, int32_t *, const WCHAR *, void **);
     using GetAsyncMethodSteppingInfoDelegate = RetCode (*)(void *, mdMethodDef, void **, int32_t *, uint32_t *);
     using GetLocalConstantsDelegate = RetCode (*)(void *, int32_t, uint32_t, void **, int32_t *);
-    using CalculationDelegate = RetCode (*)(void *, int32_t, void *, int32_t, int32_t, int32_t *, void **, BSTR *);
     using StringToUpperDelegate = RetCode (*)(const WCHAR *, BSTR *);
     using CoTaskMemFreeDelegate = void (*)(void *);
     using SysAllocStringLenDelegate = void *(*)(int32_t);
@@ -272,7 +269,6 @@ class Interop
     static CoTaskMemFreeDelegate coTaskMemFreeDelegate;
     static SysAllocStringLenDelegate sysAllocStringLenDelegate;
     static SysFreeStringDelegate sysFreeStringDelegate;
-    static CalculationDelegate calculationDelegate;
 
 };